Reafactor: Calculations moved to separate file
